code/agent_system/Captain-Agent/pyproject.toml
[project]
name = "captain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "argcomplete==1.10.3",
    "arxiv==2.3.1",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.8.2",
    "build==1.3.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "chardet==3.0.4",
    "charset-normalizer==3.4.4",
    "chromadb==1.3.4",
    "click==8.3.0",
    "cobble==0.1.4",
    "coloredlogs==15.0.1",
    "compressed-rtf==1.0.7",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docker==7.1.0",
    "docx2txt==0.9",
    "durationpy==0.10",
    "easyocr==1.7.2",
    "ebcdic==1.1.1",
    "exceptiongroup==1.3.0",
    "extract-msg==0.28.7",
    "feedparser==6.0.12",
    "filelock==3.20.0",
    "flaml==2.3.6",
    "flatbuffers==25.9.23",
    "fonttools==4.60.1",
    "fsspec==2025.10.0",
    "google-auth==2.43.0",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "imageio==2.37.2",
    "imapclient==2.1.0",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kiwisolver==1.4.9",
    "kubernetes==34.1.0",
    "lazy-loader==0.4",
    "llvmlite==0.45.1",
    "lxml==6.0.2",
    "mammoth==1.11.0",
    "markdown-it-py==4.0.0",
    "markdownify==0.6.4",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "ninja==1.13.0",
    "numba==0.62.1",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "olefile==0.47",
    "onnxruntime==1.23.2",
    "openai==2.8.0",
    "openai-whisper==20250625",
    "opencv-python-headless==4.11.0.86",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pdfminer-six==20191110",
    "pillow==12.0.0",
    "pip==23.0.1",
    "posthog==5.4.0",
    "protobuf==6.33.1",
    "py==1.11.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pyclipper==1.3.0.post6",
    "pycryptodome==3.23.0",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pymupdf==1.26.6",
    "pyparsing==3.2.5",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-bidi==0.6.7",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-pptx==0.6.23",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "retry==0.9.2",
    "rich==14.2.0",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "sentence-transformers==5.1.2",
    "setuptools==79.0.1",
    "sgmllib3k==1.0.0",
    "shapely==2.1.2",
    "shellingham==1.5.4",
    "six==1.12.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.8",
    "speechrecognition==3.8.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "termcolor==3.2.0",
    "textract==1.6.5",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tomli==2.3.0",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.3.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "wikipedia-api==0.8.1",
    "xlrd==1.2.0",
    "xlsxwriter==3.2.9",
    "zipp==3.23.0",
]
